requirements.txt
flask==3.0.0
tensorflow==2.10.0
numpy==1.26.2
opencv-python==4.8.1.78
scikit-learn==1.3.2
scipy==1.11.4
Pillow==9.5.0
google-cloud-storage
gunicorn==20.1.0